+++ tests/PythonTest.php
@@ -36,6 +36,13 @@ class PythonTest extends BaseTest {
             "{'ip_ranges': ('ip_range', ('10.0.0.1', '10.0.0.254'))}",
             "{'time_allowed': ((0, 0), (24, 0))}",
             "{'contactgroups': (True, [])}",
+            "{'socket': ('monitoring.example.com', 6559)}",
+            "{'heartbeat': (5, 2.1)}",
+            "{'key': ('abc', None)}",
+            "{'key': (True, False)}",
+            "{'key': (123, 123.45)}",
+            "{'key': ('abc_123.456-789', None)}",
+            "{'key': ((123, 456), {'a': 'b'})}",
             "{'result': {'': {}, 'foo': {'network_scan_result': {'output': 'The network scan found 0 new hosts.',
                 'state': True, 'end': 1523271725.991344, 'start': 1523271721.830718}, 'network_scan': {'time_allowed':
                 ((0, 0), (24, 0)), 'run_as': u'bar', 'tag_criticality': 'offline', 'ip_ranges': [('ip_range',
@@ -60,18 +67,31 @@ class PythonTest extends BaseTest {
                 'additional_ipv4addresses': [], u'tag_networking': u'lan', 'ipaddress': '',
                 'management_snmp_community': None}, 'configuration_hash': 'e22ba8d6831e023b771248bde84dd512'},
                 'result_code': 0}",
-            "{'result': {'site_id': 'apitestslave', 'site_config': {'url_prefix': 'http://10.10.30.230/apitestslave/',
-                'status_host': None, 'multisiteurl': 'http://10.10.30.230/apitestslave/check_mk/', 'user_sync': 'all',
-                'socket': ('proxy', {'params': None, 'socket': ('10.10.30.230', 6557)}), 'alias': u'slave',
-                'user_login': True, 'insecure': True, 'disable_wato': True, 'disabled': False, 'replication': 'slave',
-                'secret': 'ZR5=3<MG=KT04ZK853NBZ<9R4KNVE@WG', 'timeout': 10, 'persist': False, 'replicate_ec': False,
-                'replicate_mkps': True}, 'configuration_hash': '00116115ff6b27cffda0944befe72240'}, 'result_code': 0}"
+            "{'result': {'site_id': 'mysite', 'site_config': {'url_prefix': 'http://monitoring.example.com/mysite/',
+                'status_host': None, 'multisiteurl': 'http://monitoring.example.com/mysite/check_mk/',
+                'user_sync': 'all', 'socket': ('proxy', {'params': None, 'socket': ('monitoring.example.com', 6557)}),
+                'alias': u'slave', 'user_login': True, 'insecure': True, 'disable_wato': True, 'disabled': False,
+                'replication': 'slave', 'secret': 'abcdef0123456789', 'timeout': 10, 'persist': False,
+                'replicate_ec': False, 'replicate_mkps': True},
+                'configuration_hash': '00116115ff6b27cffda0944befe72240'}, 'result_code': 0}",
+            "{'result': {'site_id': 'mysite', 'site_config': {'status_host': None,
+                'url_prefix': 'http://monitoring.example.com/mysite/', 'user_sync': None,
+                'socket': ('proxy', {'socket': ('monitoring.example.com', 6559), 'query_timeout': 120.0, 'cache': True, 
+                'channels': 5, 'channel_timeout': 3.0, 'heartbeat': (5, 2.0), 'connect_retry': 4.0}),
+                'replication': 'slave', 'user_login': True, 'insecure': True, 'disable_wato': False, 'disabled': False,
+                'alias': u'My site', 'secret': 'abcdef0123456789',
+                'globals': {'cmc_cmk_helpers': 5, 'cmc_check_helpers': 2}, 'replicate_mkps': False, 'timeout': 10,
+                'persist': False, 'replicate_ec': False,
+                'multisiteurl': 'http://monitoring.example.com/mysite/check_mk/'},
+                'configuration_hash': '2f550c15bde70229bd02f20bc1bf3c5b'}, 'result_code': 0}"
         ];
 
         foreach ($values as $value) {
             $result = Python::decode($value);
 
             $this->assertInternalType('array', $result, $value);
+
+            echo json_encode($result, JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES);
         }
     }