+++ CHANGELOG.md
@@ -8,7 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ##  [Unreleased]
 
-_tbd_
+
+### Changed
+
+-   `Host::getAll()`: Fetch all "effective" attributes from hosts
 
 
 ##  [0.3] – 2018-04-25