+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   `Host::getAll()`: Fetch all "effective" attributes from hosts
 
 
+### Fixed
+
+-   Python dictionary may contain floating number in tupel
+
+
 ##  [0.3] – 2018-04-25