+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ##  [Unreleased]
 
+_tbd_
+
+
+##  [0.4]
+
 
 ### Changed
 
@@ -74,6 +79,7 @@ Initial release
 -   Get metrics
 
 
-[Unreleased]: https://github.com/bheisig/checkmkwebapi/compare/0.3...HEAD
+[Unreleased]: https://github.com/bheisig/checkmkwebapi/compare/0.4...HEAD
+[0.4]: https://github.com/bheisig/checkmkwebapi/compare/0.3...0.4
 [0.3]: https://github.com/bheisig/checkmkwebapi/compare/0.2...0.3
 [0.2]: https://github.com/bheisig/checkmkwebapi/compare/0.1...0.2